+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add version info to monitoring
 * Remove Polytimos from zpool
 * Fix displaying earnings less than 0.0001 BTC
+* Fix zero earnings when WhatToMine/pool API request fails
 
 ## RudeHash 7.1 (2018-03-20)